Consolidate etcd machine cleanup and force remove machines on etcd restore shutdown phase
