+++ pkg/capr/planner/etcdcreate.go
@@ -61,6 +61,40 @@ func (p *Planner) runEtcdSnapshotCreate(controlPlane *rkev1.RKEControlPlane, tok
 	return errs
 }
 
+// runEtcdSnapshotManagementServiceStart walks through the reconciliation process for the controlplane and etcd nodes.
+// Notably, this function will blatantly ignore drain and concurrency options, as during an etcd snapshot operation, there is no necessity to drain nodes.
+func (p *Planner) runEtcdSnapshotManagementServiceStart(controlPlane *rkev1.RKEControlPlane, tokensSecret plan.Secret, clusterPlan *plan.Plan, include roleFilter, operation string) error {
+	// Generate and deliver desired plan for the bootstrap/init node first.
+	if err := p.reconcile(controlPlane, tokensSecret, clusterPlan, true, bootstrapTier, isEtcd, isNotInitNodeOrIsDeleting,
+		"1", "",
+		controlPlane.Spec.UpgradeStrategy.ControlPlaneDrainOptions); err != nil {
+		return err
+	}
+
+	_, joinServer, _, err := p.findInitNode(controlPlane, clusterPlan)
+	if err != nil {
+		return err
+	}
+
+	if joinServer == "" {
+		return fmt.Errorf("error encountered restarting cluster during %s, joinServer was empty", operation)
+	}
+
+	for _, entry := range collect(clusterPlan, include) {
+		if isInitNodeOrDeleting(entry) {
+			continue
+		}
+		plan, joinedServer, err := p.desiredPlan(controlPlane, tokensSecret, entry, joinServer)
+		if err != nil {
+			return err
+		}
+		if err = assignAndCheckPlan(p.store, fmt.Sprintf("%s management plane restart", operation), entry, plan, joinedServer, 1, -1); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 // generateEtcdSnapshotCreatePlan generates a plan that contains an instruction to create an etcd snapshot.
 func (p *Planner) generateEtcdSnapshotCreatePlan(controlPlane *rkev1.RKEControlPlane, tokensSecret plan.Secret, entry *planEntry, joinServer string) (plan.NodePlan, string, error) {
 	v, err := semver.NewVersion(controlPlane.Spec.KubernetesVersion)
